requirements.txt
Django>=4.2.0
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-filter>=23.5
django-guardian>=2.4.0
django-rest-framework>=0.1.0
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
Pillow>=10.1.0
python-dotenv>=1.0.0
requests>=2.31.0
whitenoise>=6.6.0
djangorestframework-simplejwt>=5.3.1
drf-nested-routers>=0.93.4